STLoader: Fix bug in PCD_EP_OutXfrComplete_int() that causes USBD_LL_GetRxDataSize()/HAL_PCD_EP_GetRxCount() to return the wrong number of bytes when USB DMA mode is enabled.
Also, apparently the USB hardware only restricts writing to packet boundaries. So if you give USBD_LL_PrepareReceive()/HAL_PCD_EP_Receive() a length that isn't a multiple of the max packet size, it can write beyond the supplied buffer, up to the max packet size. So added an assertion to verify that the supplied length to HAL_PCD_EP_Receive() is a multiple of the max packet size.