ICEApp.v: attempting to remove 'SD DatOut Trigger State Machine' in favor of an ImgController 'readout_start' signal, and having SDController check 'datOutRead_ready' at the beginning of the DatOut state machine, instead of the end.
The problem with this approach is that the trailing bytes that ImgController writes to the FIFO won't be written, unless they're a multiple of the FIFO size (which they're not). So next we'll try to fix that by having ImgController guarantee that it writes a multiple of a parameterized block size.