Correct (swap) ICE SPI flash's data in/out pins so that by default, the flash is ready to be read by ICE40, but STM32 can program it by bit-banging ICE_ST_SPI_D0 / ICE_ST_SPI_D1.
(To program the SPI flash, STM32 has to bitbang instead of using the
QSPI peripheral because ICE_ST_SPI_D0 is wired between STM32's
QSPI output <-> SPI flash output, and ICE_ST_SPI_D1 is wired between
STM32's QSPI input <-> SPI flash input. That is, we have outputs
connected to outputs, and inputs connected to inputs. So for QSPI to
program the flash, the wires would need to be swapped, but that would
break the default case of ICE40 loading its configuration from SPI
flash. So we have to bitbang so that we can swap the lines to make
STM32's end of ICE_ST_SPI_D0 / ICE_ST_SPI_D1 an input / output,
respectively.)