CMD0 failures seem to be fixed by waiting longer after VDDB rails are turned on.
In short, the problem seems to be that ICE40's PLLs aren't ready when MSPApp tries to communicate with ICE40

In long, when MSPApp aborts, all rails are turned off, which means ICE40 turns off. When MSPApp turns ICE40 back on, we wait 75ms for ICE40 to turn on, and initiate comms with ICE40 via the ICE::Init() function, which succeeds. But it succeeds because it only relies on the SPI clock, not the ICE40's PLL's! During this time, ICE40 isn't fully ready yet, because the PLLs are still locking. So to tell when ICE40 is ready, we need something that depends on all the clocks in the design.