STMLoader/STMApp: detect BAT_CHRG_STAT oscillating, which we take to mean that MCP73831T is in the shutdown state (ChargeStatus::Shutdown).
For some reason BAT_CHRG_STAT oscillates on our Rev7 board, instead of remaining high-z, probably because VDD_BAT leakage current is too high for MCP73831T to remain in its shutdown state. But at least if it's oscillating, we can detect it and treat it the same way as if it were high-z.