STMApp: make USB speed and PHY tune value dynamically configurable STMLoader: enumerate in USB full-speed mode
The STMApp changes are so MDCStudio can decide the PHY tune value at runtime, depending on how Photon is connected to the Mac:

    Connection, direct to ARM Mac    -> .phyTune = 0xE043
    Connection, direct to Intel Mac  -> .phyTune = 0xF17
    Connection, via hub              -> .phyTune = 0xF17

The STMLoader changes are so the device always enumerates in a reliable configuration (USB full-speed, .phyTune = 0xF17). At least, so far we've never observed any failures with this configuration.

This is a combined commit from scratch/Dynamic-USB-Config